CovarianceUpper {fields}	R Documentation

Evaluate covariance over upper triangle of distance matrix

Description

Evaluates the covariance over the upper triangle of a distance matrix rather than over the entire matrix to reduce computation time. Note that the chol function only requires the upper triangle of the covariance matrix to perform the Cholesky decomposition.

Usage

ExponentialUpper(distMat, range = 1, alpha = 1/range)

Arguments

distMat
The distance matrix to evaluate the covariance over.

range
Range parameter default is one. Note that the scale can also be specified through the "theta" scaling argument used in fields covariance functions)

alpha
1/range

Value

The covariance matrix, where only the upper triangle is calculated.

Author(s)

John Paige

See Also

Exponential

Examples

set.seed(123)

#make distance matrix using the random locations
coords = matrix(runif(10), ncol=2)
distMat = rdist(coords)

#compute covariance matrix, but only over the upper triangle
upperCov = ExponentialUpper(distMat, range=.1)

print(distMat)
print(upperCov)

[Package fields version 9.8-5 Index]