fields exported FORTRAN {fields}	R Documentation

FORTRAN subroutines used in fields functions

Description

These functions implement cubic smoothing splines and also provide some basic computations for radial basis functions. All are called using the .FORTRAN interface.

Details

For these low level FORTRAN subroutines refer to the R functions for the calling sequence and to the src subdirectory of the fields pacakage for the source code.

css

Cubic smoothing spline see sreg and splint

ddfind

Finds nearest neighbor points within a fixed distance. See fields.rdist.near

dmaket

Creates matrix of all polynomial terms up to fixed order. See fields.mkpoly

evlpoly

evaluates a univariate polynomial. See codefields.evlpoly

evlpoly2

evaluates a multivariate polynomial. See codefields.evlpoly2

inpoly

Determine which 2-d locations are within a polynomial. see in.poly

igpoly

Determine which 2-d grid points locations are within a polynomial. see in.poly.grid

multeb

Multiply an exponential cross covariance matrix by another matrix. See exp.cov

multrb

Multiply an radial basis function matrix by another matrix. See rad.cov

radbas

Evaluates radial basis functions. See rdist.R

rcss

Robust cubic smoothing spline. See qsreg

[Package fields version 9.6 Index]