Krig.null.function {fields}	R Documentation

Default function to create fixed matrix part of spatial process model.

Description

Constructs a matrix of terms representing a low order polynomial and binds additional columns due to covariates ( the Z matrix)

Usage

Krig.null.function(x, Z = NULL, drop.Z = FALSE, m)

Arguments

x
Spatial locations

Z
Other covariates to be associated with each location.

drop.Z
If TRUE only the low order polynomial part is created.

m
The polynomial order is (m-1).

Details

This function can be modified to produce a different fixed part of the spatial model. The arguments x, Z and drop.Z are required but other arguments can be passed as part of a list in null.args in the call to Krig.

Value

A matrix where the first columns are the polynomial terms and the following columns are from Z.

Author(s)

Doug Nychka

See Also

Krig

[Package fields version 9.9 Index]