US {fields}	R Documentation

Plot of the US with state boundaries

Description

Plots quickly, medium resolution outlines of the US with the states and bodies of water. A simple wrapper for the map function from the maps package.

Usage

US( ...)

Arguments

...
These are the arguments that are passed to the map function from the maps package.

Details

The older version of this function (fields < 6.7.2) used the FIELDS dataset US.dat for the coordinates. Currenty this has been switched to use the maps package.

See Also

world

Examples

# Draw map in device color # 3 
US( col=3) 

[Package fields version 12.6 Index]