US.dat {fields}	R Documentation

Outline of coterminous US and states.

Description

This data set is used by the fields function US to draw a map. It is the medium resolution outline that is produced by drawing the US from the maps package.

[Package fields version 12.6 Index]