add.image {fields}	R Documentation

Adds an image to an existing plot.

Description

Adds an image to an existing plot. Simple arguments control the location and size.

Usage

add.image(xpos, ypos, z, adj.x = 0.5, adj.y = 0.5, 
image.width = 0.15, image.height = NULL, col = tim.colors(256), ...)

Arguments

xpos
X position of image in user coordinates

ypos
Y position of image in user coordinates

z
Matrix of intensities comprising the image.

adj.x
Location of image relative to x coordinate. Most common values are .5 (centered), 0 (right side of image at x) and 1 (left side of image at x). These are the same conventions that are used for adj in positioning text.

adj.y
Location of image relative to y coordinate. Same rules as adj.x

image.width
Width of image as a fraction of the plotting region in horizontal direction.

image.height
Height of image as a fraction of the plotting region in horizontal direction. If NULL height is scaled to make image pixels square.

col
Color table for image. Default is tim.colors.

...
Any other plotting arguments that are passed to the image function

See Also

image.plot, colorbar.plot, image, tim.colors

Examples

plot( 1:10, 1:10, type="n")
data( lennon)

add.image( 5,4,lennon, col=grey( (0:256)/256))
# reference lines 
xline( 5, col=2)
yline( 4,col=2) 

#
# add lennon right in the corner beyond the plotting region
# 

par(new=TRUE, plt=c(0,1,0,1), mar=c(0,0,0,0), usr=c(0,1,0,1))
add.image( 0,0, lennon, adj.x=0, adj.y=0) 

[Package fields version 9.8-5 Index]