flame {fields}	R Documentation

Response surface experiment ionizing a reagent

Description

The characteristics of an ionizing flame are varied with the intent of maximizing the intensity of emitted light for lithuim in solution. Areas outside of the measurements are where the mixture may explode! Note that the optimum is close to the boundary. Source of data is from a master's level lab experiment in analytical chemistry from Chuck Boss's course at NCSU. <s-section name= "DATA DESCRIPTION"> This is list with the following components

Arguments

x
x is a 2 column matrix with the different Fuel and oxygen flow rates for the burner.

y
y is the response. The intensity of light at a particular wavelength indicative of Lithium ions.

[Package fields version 9.8-5 Index]