minitri {fields}	R Documentation

Mini triathlon results

Description

Results from a mini triathlon sponsored by Bud Lite, held in Cary, NC, June 1990. Times are in minutes for the male 30-34 group. Man was it hot and humid! (DN)

The events in order were swim: (1/2 mile) bike: (15 miles) run: (4 miles)

<s-section name= "DATA DESCRIPTION"> This is a dataframe. Row names are the place within this age group based on total time.

Arguments

swim
swim times

bike
bike times

run
run times

[Package fields version 9.8-5 Index]