plot.Krig {fields}	R Documentation

Diagnostic and summary plots of a Kriging, spatialProcess or spline object.

Description

Plots a series of four diagnostic plots that summarize the fit.

Usage

## S3 method for class 'Krig'
plot(x, digits=4, which= 1:4,...)
## S3 method for class 'sreg'
plot(x, digits = 4, which = 1:4, ...)

Arguments

x
A Krig or an sreg object

digits
Number of significant digits for the RMSE label.

which
A vector specifying by number which of the four plots to draw. 1:4 plots all four.

...
Optional graphics arguments to pass to each plot.

Details

This function creates four summary plots of the Krig or sreg object. The default is to put these on separate pages. However if the screen is already divided in some other fashion the plots will just be added according to that scheme. This option is useful to compare to compare several different model fits.

The first is a scatterplot of predicted value against observed.

The second plot is "standardized" residuals against predicted value. Here we mean that the residuals are divided by the GCV estimate for sigma and multiplied by the square root of any weights that have been specified. In the case of a "correlation model" the residuals are also divided by the marginal standard deviation from this model.

The third plot are the values of the GCV function against the effective degrees of freedom. When there are replicate points several versions of the GCV function may be plotted. GCV function is with respect to the standardized data if a correlation model is specified. A vertical line indicates the minimium found.

For Krig and sreg objects the fourth plot is a histogram of the standardized residuals. For sreg if multiple lambdas are given plotted are boxplots of the residuals for each fit.

For spatialProcess object the fourth plot is the profile likelihood for the theta parameter. Points are the actual evaluated log likelihoods and the dashed line is just a spline interpolation to help with visualization.

See Also

Krig, spatialProcess, summary.Krig, Tps, set.panel

Examples

data( ozone2)
x<- ozone2$lon.lat
y<- ozone2$y[16,]
fit1<-Krig(x,y, theta=200)  
# fitting a surface to ozone  
# measurements 
set.panel( 2,2)
plot(fit1)

fit2<-spatialProcess(x,y)  
# fitting a spatial process model to ozone  
# measurements 
# Although an example does not make too much sense for only 20 observations!
set.panel( 2,2)
plot(fit2)

# fit rat data
fit3<-sreg(rat.diet$t,rat.diet$con)
set.panel(2,2)
plot(fit3)       

set.panel(1,1) # reset graphics window. 

[Package fields version 9.9 Index]