pushpin {fields}	R Documentation

Adds a "push pin" to an existing 3-d plot

Description

Adds to an existing 3-d perspective plot a push pin to locate a specific point.

Usage

pushpin( x,y,z,p.out, height=.05,col="black",text=NULL,adj=-.1,cex=1.0,...)

Arguments

x
x location

y
y location

z
z location

p.out
Projection information returned by persp

height
Height of pin in device coordinates (default is about 5% of the vertical distance ).

col
Color of pin head.

text
Optional text to go next to pin head.

adj
Position of text relative to pin head.

cex
Character size for pin head and/or text

...
Additional graphics arguments that are passed to the text function.

Details

See the help(text) for the conventions on the adj argument and other options for placing text.

Author(s)

Doug Nychka

See Also

drape.plot,persp

Examples

# Dr. R's favorite New  Zealand Volcano!
     data( volcano)
     M<- nrow( volcano)
     N<- ncol( volcano)
     x<- seq( 0,1,,M)
     y<- seq( 0,1,,N)

     drape.plot( x,y,volcano, col=terrain.colors(128))-> pm 

     max( volcano)-> zsummit
     xsummit<- x[ row( volcano)[volcano==zsummit]]
     ysummit<- y[ col( volcano)[volcano==zsummit]]

pushpin( xsummit,ysummit,zsummit,pm, text="Summit")

[Package fields version 12.6 Index]