rat.diet {fields}	R Documentation

Experiment studying an appetite supressant in rats.

Description

The ‘rat.diet’ data frame has 39 rows and 3 columns. These are data from a study of an appetite supressant given to young rats. The suppressant was removed from the treatment group at around 60 days. The responses are the median food intake and each group had approximately 10 animals.

Usage

data(rat.diet)

Format

This data frame contains the following columns:

t

Time in days

con

Median food intake of the control group

trt

Median food intake of the treatment group

[Package fields version 9.8-5 Index]