Information objects that register C and FORTRAN functions.

These are objects of class CallRoutine or FortranRoutine and also NativeSymbolInfo They provide information for compiledfunctions called with .Call, or .Fortran. Ordinarily one would not need to consult these and they are used to make the search among dynamically loaded libraries ( in particular the fields library) have less ambiguity and also be faster. These are created when the package/library is loaded are have their definitions from the compliation of init.c in the package source (src) directory.

Format

The format is a list with components:

name

The (registration ?) name of the C function.

address

See NativeSymbolInfo.

dll

Dynamically linked library information.

numParameters

Number of calling arguments in function.

Details

addToDiagC

adds diagonal elements to a matrix. See codemKrig.

ExponentialUpperC

Fills in upper triangle of a matrix with the exponential covariance function. See ExponentialUpper

compactToMatC

Converts compact format to full matrix format. See compactToMat.

multebC

Mulitplies a vector/matrix with an exponential covariance function. See exp.cov

multwendlandg

This has been mysteriously included but it is not a function!

mltdrb

Evaluates the derivatives of thin plate sline radial basis functions. See rad.cov.

RdistC

Euclidean distance function between sets of coordinates. See rdist.

distMatHaversin

Used in RdistEarth.

distMatHaversin2

Used in RdistEarth.

See package_native_routine_registration_skeleton for the utility used to create these data objects.

It is not clear why these routines have been flagged as needing documentation while other routines have not.

References

For background on registering C, C++ and Fortran functions see 5.4 of Writing R Extensions. For this package refer to the C code in src/intit.c as an example.

Examples

print(addToDiagC)
#> $name
#> [1] "addToDiagC"
#> 
#> $address
#> <pointer: 0x55a9f1a35cd0>
#> attr(,"class")
#> [1] "RegisteredNativeSymbol"
#> 
#> $dll
#> DLL name: fields
#> Filename: /home/runner/work/_temp/Library/fields/libs/fields.so
#> Dynamic lookup: FALSE
#> 
#> $numParameters
#> [1] 3
#> 
#> attr(,"class")
#> [1] "CallRoutine"      "NativeSymbolInfo"