Summary for Krig or spatialProcess estimated models.

Creates a list of summary results including estimates for the nugget variance (tau) and the smoothing parameter (lambda). This list is usually printed using a "print.summary" function for nice formatting.

# S3 method for Krig
summary(object, digits=4,...)

Arguments

object

A Krig or spatialProcess object.

digits

Number of significant digits in summary.

...

Other arguments to summary

Value

Gives a summary of the Krig object. The components include the function call, number of observations, effective degrees of freedom, residual degrees of freedom, root mean squared error, R-squared and adjusted R-squared, log10(lambda), cost, GCV minimum and a summary of the residuals.

Details

This function is a method for the generic function summary for class Krig. The results are formatted and printed using print.summary.Krig.

See also

Examples

fit<- Krig(ChicagoO3$x, ChicagoO3$y, aRange=100)  
summary(fit)                            # summary of fit 
#> CALL:
#> Krig(x = ChicagoO3$x, Y = ChicagoO3$y, aRange = 100)
#>                                               
#>  Number of Observations:                20    
#>  Number of unique points:               20    
#>  Number of parameters in the null space 3     
#>  Parameters for fixed spatial drift     3     
#>  Effective degrees of freedom:          5.4   
#>  Residual degrees of freedom:           14.6  
#>  MLE tau                                3.699 
#>  GCV tau                                4.012 
#>  MLE sigma                              20.25 
#>  Scale passed for covariance (sigma)    <NA>  
#>  Scale passed for nugget (tau^2)        <NA>  
#>  Smoothing parameter lambda             0.6755
#> 
#> Residual Summary:
#>     min   1st Q  median   3rd Q     max 
#> -6.3880 -1.4160 -0.5873  1.5540  7.5930 
#> 
#> Covariance Model: stationary.cov
#>   Covariance function is 
#>   Names of non-default covariance arguments: 
#>        aRange
#> 
#> DETAILS ON SMOOTHING PARAMETER:
#>  Method used:   REML    Cost:  1
#>    lambda       trA       GCV   GCV.one GCV.model    tauHat 
#>    0.6755    5.4421   22.1105   22.1105        NA    4.0117 
#> 
#>  Summary of all estimates found for lambda
#>            lambda   trA   GCV tauHat -lnLike Prof converge
#> GCV        0.9654 4.842 22.02  4.086        49.16        4
#> GCV.model      NA    NA    NA     NA           NA       NA
#> GCV.one    0.9654 4.842 22.02  4.086           NA        4
#> RMSE           NA    NA    NA     NA           NA       NA
#> pure error     NA    NA    NA     NA           NA       NA
#> REML       0.6755 5.442 22.11  4.012        49.15        3