summary.ncdf {fields}	R Documentation

Summarizes a netCDF file handle

Description

Provides a summary of the variable names and sizes from the handle returned from netCDF file.

Usage

## S3 method for class 'ncdf'
summary(object,...)

Arguments

object
The "handle" returned by the read.ncdf function from the ncdf package.

...
Other arguments to pass to this function. Currently, no other arguments are used.

Details

This function is out of place in fields but was included because often large geophysical data sets are in netCDF format and the ncdf R package is also needed. To date the summary capability in the ncdf package is limited and this function is used to supplement it use. The function is also a a useful device to see how the ncdf object is structured.

Author(s)

D. Nychka

See Also

ncdf

[Package fields version 12.6 Index]