transformx {fields}	R Documentation

Linear transformation

Description

Linear transformation of each column of a matrix. There are several choices of the type of centering and scaling.

Usage

transformx (x, scale.type = "unit.sd", x.center, x.scale)

Arguments

x
Matrix with columns to be transformed.

scale.type
Type of transformation the default is "unit.sd": subtract the mean and divide by the standard deviation. Other choices are "unscaled" (do nothing), "range" (transform to [0,1]),"user" (subtract a supplied location and divide by a scale).

x.center
A vector of centering values to subtract from each column.

x.scale
A vector of scaling values to subtract from each column.

Details

After deciding what the centering and scaling values should be for each column of x, this function just calls the standard utility scale. This function was created partly to attach the transformation information as attributes to the transformed matrix. It is used in Krig, cover.design, krig.image etc. to transform the independent variables.

Value

A matrix whose columns have between transformed. This matrix also has the attributes: scale.type, x.center and y.center with the transformation information.

See Also

scale

Examples

#
newx<-transformx( ChicagoO3$x, scale.type="range")

[Package fields version 9.9 Index]