world {fields}	R Documentation

Plot of the world

Description

Plots quickly, medium resolution outlines of large land masses. This is a simple wrapper for the map function from the maps package.

Usage

world(...)
world.land(  ...)
world.color( ... )
in.land.grid(...)

Arguments

...
Same arguments used by the map function from the maps package.

Details

See the longstanding maps package for documentation on this function. The functions world.land, world.color and in.land.grid have been depreciated but can be recovered from versions of fields 6.7.1 or older.

See Also

US, in.poly, in.poly.grid

Examples

## Not run: 
world()
# add the US
US( add=TRUE,col="blue")

world( fill=TRUE) # land filled in black 

## Western Europe
world( xlim=c(-10,18),ylim=c(36,60),fill=TRUE, col="darkgreen",
border="green1")

## End(Not run)

[Package fields version 9.9 Index]