summary.Krig {fields}	R Documentation

Summary for Krig or spatialProcess estimated models.

Description

Creates a list of summary results including estimates for the nugget variance (sigma) and the smoothing parameter (lambda). This list is usually printed using a "print.summary" function for nice formatting.

Usage

## S3 method for class 'Krig'
summary(object, digits=4,...)

Arguments

object
A Krig or spatialProcess object.

digits
Number of significant digits in summary.

...
Other arguments to summary

Details

This function is a method for the generic function summary for class Krig. The results are formatted and printed using print.summary.Krig.

Value

Gives a summary of the Krig object. The components include the function call, number of observations, effective degrees of freedom, residual degrees of freedom, root mean squared error, R-squared and adjusted R-squared, log10(lambda), cost, GCV minimum and a summary of the residuals.

See Also

Krig, summary, print.summary.Krig, summary.spatialProcess

Examples

fit<- Krig(ChicagoO3$x, ChicagoO3$y, theta=100)  
summary(fit)                            # summary of fit 

[Package fields version 9.8-5 Index]